* Sat Jan 21 2012 Ankur Sinha <ankursinha AT fedoraproject DOT org>   completed the objects I think all the objects are now complete. I need to write a main method and act according to the passed command line arguments.
I also need to add support for the config file, although I don't think
that's really tough using the program_options library

It is printing the final png well. I need to add code to iterate over
the ticket list and print them. Again, that should probably be in the
main method. I could add a method for that in ExportData too. I'll
think about it.